deleted old recyclerview stuff
