+++ DailyJournalApp/app/src/main/java/com/example/dailyjournalapp/JournalListFragment.java
@@ -3,6 +3,9 @@ package com.example.dailyjournalapp;
 import android.database.Cursor;
 import android.os.Bundle;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ListView;
@@ -20,6 +23,7 @@ import java.util.Map;
 public class JournalListFragment extends Fragment {
 
     public JournalListFragment () {}
+
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -30,6 +34,10 @@ public class JournalListFragment extends Fragment {
 
         View rootView = inflater.inflate(R.layout.fragment_list, container, false);
 
+        // enable menu
+        setHasOptionsMenu(true);
+
+        // begin adapter process
         Cursor mCursor = getActivity().getContentResolver()
                 .query(JournalContentProvider.CONTENT_URI, null, null, null, null);
 
@@ -48,11 +56,43 @@ public class JournalListFragment extends Fragment {
         int[] to = new int[] {android.R.id.text1};
 
         SimpleAdapter adapter = new SimpleAdapter(getActivity(), list, android.R.layout.simple_list_item_1, from, to);
+        adapter.notifyDataSetChanged();
 
+        // setting adapter
         ListView lv = (ListView) rootView.findViewById(R.id.list);
         lv.setAdapter(adapter);
 
+        // TODO: update adapter?
+
         return rootView;
+    }
+
+    @Override
+    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+        inflater.inflate(R.menu.menu, menu);
+        return;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case R.id.opt_set_notif:
+                NotificationFragment notifFrag = new NotificationFragment();
+                getActivity().getSupportFragmentManager().beginTransaction()
+                        .replace(R.id.frameLayout, notifFrag, "notification_fragment")
+                        .addToBackStack(null)
+                        .commit();
+                break;
+            case R.id.opt_new_entry:
+                EditJournalFragment editFrag = new EditJournalFragment();
+                getActivity().getSupportFragmentManager().beginTransaction()
+                        .replace(R.id.frameLayout, editFrag, "edit_journal_fragment")
+                        .addToBackStack(null)
+                        .commit();
+                break;
+
+        }
+        return true;
     }
 
 }
