--- DailyJournalApp/app/src/main/java/com/example/dailyjournalapp/MyJournalItemRecyclerViewAdapter.java
@@ -1,74 +1,0 @@
-package com.example.dailyjournalapp;
-
-import androidx.recyclerview.widget.RecyclerView;
-
-import android.database.Cursor;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import java.util.List;
-import java.util.Map;
-
-public class MyJournalItemRecyclerViewAdapter extends RecyclerView.Adapter<MyJournalItemRecyclerViewAdapter.ViewHolder> {
-
-    private final List<Map<Integer, String>> mValues;
-    private final OnItemClickListener mListener;
-
-    public interface OnItemClickListener {
-        void onItemClick(Map<Integer, String> item);
-    }
-
-    public MyJournalItemRecyclerViewAdapter(List<Map<Integer, String>> items, OnItemClickListener listener) {
-        mListener = listener;
-        mValues = items;
-    }
-
-    @Override
-    public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext())
-                .inflate(R.layout.fragment_item, parent, false);
-        return new ViewHolder(view);
-    }
-
-    @Override
-    public void onBindViewHolder(final ViewHolder holder, int position) {
-        holder.mItem = mValues.get(position);
-        holder.mIdView.setText(mValues.get(position).get(0));
-        holder.mContentView.setText(mValues.get(position).get(1));
-        holder.bind(mValues.get(position), mListener);
-    }
-
-    @Override
-    public int getItemCount() {
-        return mValues.size();
-    }
-
-    public class ViewHolder extends RecyclerView.ViewHolder {
-        public final View mView;
-        public final TextView mIdView;
-        public final TextView mContentView;
-        public Map<Integer, String> mItem;
-
-        public ViewHolder(View view) {
-            super(view);
-            mView = view;
-            mIdView = (TextView) view.findViewById(R.id.item_number);
-            mContentView = (TextView) view.findViewById(R.id.content);
-        }
-
-        public void bind(final Map<Integer, String> item, final OnItemClickListener listener) {
-            itemView.setOnClickListener(new View.OnClickListener() {
-                @Override public void onClick(View v) {
-                    listener.onItemClick(item);
-                }
-            });
-        }
-
-        @Override
-        public String toString() {
-            return super.toString() + " '" + mContentView.getText() + "'";
-        }
-    }
-}
